fixed comments from code review
