+++ src/mocks/localStorage.mock.js
@@ -1,0 +1,21 @@
+export class LocalStorageMock {
+    constructor() {
+      this.store = {};
+    }
+  
+    clear() {
+      this.store = {};
+    }
+  
+    getItem(key) {
+      return this.store[key] || null;
+    }
+  
+    setItem(key, value) {
+      this.store[key] = value.toString();
+    }
+  
+    removeItem(key) {
+      delete this.store[key];
+    }
+  }
