+++ src/test/storage.test.js
@@ -1,30 +1,8 @@
 import * as storage from '../js/storage';
-
-class LocalStorageMock {
-  constructor() {
-    this.store = {};
-  }
-
-  clear() {
-    this.store = {};
-  }
-
-  getItem(key) {
-    return this.store[key] || null;
-  }
-
-  setItem(key, value) {
-    this.store[key] = value.toString();
-  }
-
-  removeItem(key) {
-    delete this.store[key];
-  }
-}
-
-globalThis.localStorage = new LocalStorageMock();
+import { LocalStorageMock } from '../mocks/localStorage.mock';
 
 describe('storage', () => {
+  globalThis.localStorage = new LocalStorageMock();
   it('Saves an array to storage', () => {
     const key = 'basket';
     const value = ['item1', 'item2'];
